voice-lingo/backend/requirements.txt
'''fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.3.5
openai==1.3.5
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0
bcrypt==4.1.1
passlib[bcrypt]==1.7.4'''
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.5
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0
bcrypt==4.1.1
httpx==0.25.2
numpy==1.26.2